feat: Move auth package to `cloudquery-api-go`
Updates to the latest `cloudquery-api-go` version

fixes: https://github.com/cloudquery/cloudquery-issues/issues/668
