+++ cli/cmd/login.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/cloudquery/cloudquery-api-go/auth"
+	"github.com/cloudquery/cloudquery-api-go/config"
 	"io"
 	"net"
 	"net/http"
@@ -14,8 +16,7 @@ import (
 	"time"
 
 	"github.com/cenkalti/backoff/v4"
-	"github.com/cloudquery/cloudquery/cli/internal/auth"
-	"github.com/cloudquery/cloudquery/cli/internal/config"
+	"github.com/cloudquery/cloudquery/cli/internal/team"
 	"github.com/pkg/browser"
 	"github.com/spf13/cobra"
 )
@@ -160,20 +161,21 @@ func runLogin(ctx context.Context, cmd *cobra.Command) (err error) {
 	}
 
 	if cmd.Flags().Changed("team") {
-		team := cmd.Flag("team").Value.String()
-		token, err := auth.GetToken()
+		selectedTeam := cmd.Flag("team").Value.String()
+		tc := auth.NewTokenClient()
+		token, err := tc.GetToken()
 		if err != nil {
-			return fmt.Errorf("failed to get auth token: %w", err)
+			return fmt.Errorf("failed to get team token: %w", err)
 		}
-		cl, err := auth.NewClient(apiURL, token)
+		cl, err := team.NewClient(apiURL, token)
 		if err != nil {
 			return fmt.Errorf("failed to create API client: %w", err)
 		}
-		err = cl.ValidateTeam(ctx, team)
+		err = cl.ValidateTeam(ctx, selectedTeam)
 		if err != nil {
 			return fmt.Errorf("failed to set team: %w", err)
 		}
-		err = config.SetValue("team", team)
+		err = config.SetValue("team", selectedTeam)
 		if err != nil {
 			return fmt.Errorf("failed to set team: %w", err)
 		}
