+++ cli/cmd/switch.go
@@ -3,11 +3,12 @@ package cmd
 import (
 	"errors"
 	"fmt"
+	"github.com/cloudquery/cloudquery-api-go/auth"
+	"github.com/cloudquery/cloudquery-api-go/config"
 	"os"
 	"strings"
 
-	"github.com/cloudquery/cloudquery/cli/internal/auth"
-	"github.com/cloudquery/cloudquery/cli/internal/config"
+	"github.com/cloudquery/cloudquery/cli/internal/team"
 	"github.com/spf13/cobra"
 )
 
@@ -36,12 +37,13 @@ func newCmdSwitch() *cobra.Command {
 func runSwitch(cmd *cobra.Command, args []string) error {
 	apiURL := getEnvOrDefault("CLOUDQUERY_API_URL", defaultAPIURL)
 
-	token, err := auth.GetToken()
+	tc := auth.NewTokenClient()
+	token, err := tc.GetToken()
 	if err != nil {
-		return fmt.Errorf("failed to get auth token: %w", err)
+		return fmt.Errorf("failed to get team token: %w", err)
 	}
 
-	cl, err := auth.NewClient(apiURL, token)
+	cl, err := team.NewClient(apiURL, token)
 	if err != nil {
 		return fmt.Errorf("failed to create API client: %w", err)
 	}
@@ -70,15 +72,15 @@ func runSwitch(cmd *cobra.Command, args []string) error {
 		cmd.Println("To switch teams, run `cloudquery switch <team>`")
 		return nil
 	}
-	team := args[0]
-	err = cl.ValidateTeam(cmd.Context(), team)
+	selectedTeam := args[0]
+	err = cl.ValidateTeam(cmd.Context(), selectedTeam)
 	if err != nil {
 		return fmt.Errorf("failed to switch teams: %w", err)
 	}
-	err = config.SetValue("team", team)
+	err = config.SetValue("team", selectedTeam)
 	if err != nil {
 		return fmt.Errorf("failed to set team value: %w", err)
 	}
-	cmd.Printf("Successfully switched teams to %v.\n", team)
+	cmd.Printf("Successfully switched teams to %v.\n", selectedTeam)
 	return nil
 }
