+++ cli/internal/team/logout.go
@@ -1,13 +1,13 @@
-package auth
+package team
 
 import (
 	"fmt"
-
-	"github.com/cloudquery/cloudquery/cli/internal/config"
+	"github.com/cloudquery/cloudquery-api-go/auth"
+	"github.com/cloudquery/cloudquery-api-go/config"
 )
 
 func Logout() error {
-	err := removeRefreshToken()
+	err := auth.RemoveRefreshToken()
 	if err != nil {
 		return fmt.Errorf("failed to remove refresh token: %w", err)
 	}
